Add files via upload
The aim of this project was to develop a deep learning model that can accurately classify brain tumors into three categories: Meningioma, Pituitary, or Glioma. I used Resnet50 as a pre-trained model and fine-tuned it on a dataset of MRI images of brain tumors. The model achieved a high accuracy of 97.5% on the test set, demonstrating its effectiveness and robustness for this task. After doing K-Fold cross validation on the I achieved an average accuracy of 96.84% with a standard deviation of 0.39%.